safari-534.53-branch/LayoutTests/svg/dynamic-updates/SVGFEConvolveMatrixElement-svgdom-kernelMatrix-prop-expected.txt

CONSOLE MESSAGE: line 55: TypeError: 'undefined' is not an object (evaluating 'convolveMatrixElement.kernelMatrix.baseVal')
FAIL: Timed out waiting for notifyDone to be called
SVG 1.1 dynamic update tests

Tests dynamic updates of the 'kernelMatrix' property of the SVGFEConvolveMatrixElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL SVGNumberListToString(convolveMatrixElement.kernelMatrix.baseVal) should be -2 0 0 0 1 0 0 0 2 . Threw exception TypeError: 'undefined' is not an object (evaluating 'convolveMatrixElement.kernelMatrix.baseVal')
